Update README.md
Update the description of the Dart ic_tools library to clarify that the library can run on Dart code by itself and on a linux desktop. 

Add the ic_tools_web library which contains specific tools and classes for the Flutter-Web platform such as a Caller class using the SubtleCrypto export=false browser api, and an internet-identity login function.