Add AI chatbot DeVinci
I built this chat app prototype around the open-source project WebLLM and deployed it on the Internet Computer (so the frontend is served from the IC and the backend canister stores the user's data, especially the chats here). This way, we've got a fully decentralized AI chatbot app that runs on the user's device (via the browser). If you like, this PR adds it as an example. Thank you.